✨ feat: recommendations 서비스, 리포지토리 클래스 분리
