+++ app/src/main/kotlin/com/firstpenguin/app/domain/recommendation/useCase/RecommendationUseCase.kt
@@ -1,19 +1,17 @@
 package com.firstpenguin.app.domain.recommendation.useCase
 
-import com.firstpenguin.app.domain.book.service.BookService
-import com.firstpenguin.app.domain.emotion.dto.TagDto
-import com.firstpenguin.app.domain.emotion.service.EmotionService
 import com.firstpenguin.app.domain.quote.dto.QuoteResponse
-import com.firstpenguin.app.domain.quote.model.Quote
 import com.firstpenguin.app.domain.quote.service.QuoteService
 import com.firstpenguin.app.domain.recommendation.dto.RecommendationDetailResponse
 import com.firstpenguin.app.domain.recommendation.dto.RecommendationPeriodItemResponse
 import com.firstpenguin.app.domain.recommendation.dto.RecommendationPeriodResponse
 import com.firstpenguin.app.domain.recommendation.dto.RecommendationRequest
 import com.firstpenguin.app.domain.recommendation.dto.RecommendationResponse
-import com.firstpenguin.app.domain.recommendation.model.Recommendation
-import com.firstpenguin.app.domain.recommendation.model.RecommendationTag
+import com.firstpenguin.app.domain.recommendation.service.RecommendationCommandService
+import com.firstpenguin.app.domain.recommendation.service.RecommendationRequestValidator
+import com.firstpenguin.app.domain.recommendation.service.RecommendationResponseMapper
 import com.firstpenguin.app.domain.recommendation.service.RecommendationService
+import com.firstpenguin.app.domain.recommendation.service.RecommendationValidationService
 import com.firstpenguin.app.global.exception.CustomException
 import com.firstpenguin.app.global.exception.ErrorCode
 import org.springframework.stereotype.Service
@@ -21,36 +19,31 @@ import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDate
 
 private const val NEXT_RECOMMENDATION_QUOTE_COUNT = 9
-private const val MAX_EMOTION_TAG_COUNT = 5
 
 @Service
 class RecommendationUseCase(
-    private val emotionService: EmotionService,
     private val quoteService: QuoteService,
     private val recommendationService: RecommendationService,
-    private val bookService: BookService,
+    private val recommendationCommandService: RecommendationCommandService,
+    private val recommendationValidationService: RecommendationValidationService,
+    private val recommendationRequestValidator: RecommendationRequestValidator,
+    private val recommendationResponseMapper: RecommendationResponseMapper,
 ) {
     @Transactional
     fun recommendQuote(
         userId: Long,
         request: RecommendationRequest,
     ): RecommendationResponse {
-        validateRecommendationRequest(request)
+        recommendationRequestValidator.validate(request)
 
         val selectedTagIds = request.emotionTagIds + listOfNotNull(request.needTagId)
 
-        emotionService.validateTags(
-            emotionRangeId = request.emotionRangeId,
-            emotionTagIds = request.emotionTagIds,
-            needTagId = request.needTagId,
-        )
-
         recommendationService.lockRecommendationCreation(userId)
         recommendationService.findOngoingRecommendation(userId)?.let { recommendation ->
-            return toRecommendationResponse(recommendation)
+            return recommendationResponseMapper.toRecommendationResponse(recommendation)
         }
 
-        recommendationService.validateRecommendationCreatable(userId)
+        recommendationValidationService.validateRecommendationCreatable(userId)
 
         val recommendationId =
             recommendationService.createRecommendation(
@@ -70,30 +63,10 @@ class RecommendationUseCase(
             quoteIds = listOf(randomQuote.id),
         )
 
-        return toRecommendationResponse(
+        return recommendationResponseMapper.toRecommendationResponse(
             recommendationId = recommendationId,
             quote = randomQuote,
         )
-    }
-
-    private fun validateRecommendationRequest(request: RecommendationRequest) {
-        validateEmotionTagCount(request.emotionTagIds)
-        validateNeedInput(request)
-    }
-
-    private fun validateEmotionTagCount(emotionTagIds: List<Long>) {
-        if (emotionTagIds.isEmpty() || emotionTagIds.size > MAX_EMOTION_TAG_COUNT) {
-            throw CustomException(ErrorCode.INVALID_EMOTION_TAG)
-        }
-    }
-
-    private fun validateNeedInput(request: RecommendationRequest) {
-        val hasNeedTag = request.needTagId != null
-        val hasFeelingText = request.feelingText.hasText()
-
-        if (hasNeedTag == hasFeelingText) {
-            throw CustomException(ErrorCode.INVALID_RECOMMENDATION_NEED_INPUT)
-        }
     }
 
     @Transactional
@@ -101,11 +74,11 @@ class RecommendationUseCase(
         userId: Long,
         recommendationId: Long,
     ): List<QuoteResponse> {
-        val recommendation = recommendationService.validateRecommendation(userId, recommendationId)
-        recommendationService.validateRecommendationOngoing(recommendation)
+        val recommendation = recommendationValidationService.validateRecommendation(userId, recommendationId)
+        recommendationValidationService.validateRecommendationOngoing(recommendation)
 
         val recommendationHistory = recommendationService.getRecommendationHistory(recommendationId)
-        recommendationService.validateRecommendationCount(
+        recommendationValidationService.validateRecommendationCount(
             currentCount = recommendationHistory.size,
             nextCount = NEXT_RECOMMENDATION_QUOTE_COUNT,
         )
@@ -122,7 +95,7 @@ class RecommendationUseCase(
             quoteIds = nextQuotes.map { quote -> quote.id },
         )
 
-        return nextQuotes.map(::toQuoteResponse)
+        return nextQuotes.map(recommendationResponseMapper::toQuoteResponse)
     }
 
     @Transactional
@@ -131,8 +104,8 @@ class RecommendationUseCase(
         recommendationId: Long,
         quoteId: Long,
     ): RecommendationDetailResponse =
-        toRecommendationDetailResponse(
-            recommendationService.selectRecommendationQuote(
+        recommendationResponseMapper.toRecommendationDetailResponse(
+            recommendationCommandService.selectRecommendationQuote(
                 userId = userId,
                 recommendationId = recommendationId,
                 quoteId = quoteId,
@@ -144,10 +117,10 @@ class RecommendationUseCase(
         userId: Long,
         recommendationId: Long,
     ) {
-        val recommendation = recommendationService.validateRecommendation(userId, recommendationId)
-        recommendationService.validateRecommendationCompleted(recommendation)
+        val recommendation = recommendationValidationService.validateRecommendation(userId, recommendationId)
+        recommendationValidationService.validateRecommendationCompleted(recommendation)
 
-        recommendationService.deleteRecommendation(recommendationId)
+        recommendationCommandService.deleteRecommendation(recommendationId)
     }
 
     @Transactional(readOnly = true)
@@ -156,13 +129,14 @@ class RecommendationUseCase(
         recommendationId: Long,
     ): RecommendationDetailResponse {
         val recommendation =
-            recommendationService.validateRecommendation(
+            recommendationValidationService.validateRecommendation(
                 userId = userId,
                 recommendationId = recommendationId,
                 lockForUpdate = false,
             )
+        recommendationValidationService.validateRecommendationCompleted(recommendation)
 
-        return toRecommendationDetailResponse(recommendation)
+        return recommendationResponseMapper.toRecommendationDetailResponse(recommendation)
     }
 
     @Transactional(readOnly = true)
@@ -170,7 +144,7 @@ class RecommendationUseCase(
         userId: Long,
         recommendationId: Long,
     ): List<QuoteResponse> {
-        recommendationService.validateRecommendation(
+        recommendationValidationService.validateRecommendation(
             userId = userId,
             recommendationId = recommendationId,
             lockForUpdate = false,
@@ -179,7 +153,7 @@ class RecommendationUseCase(
         return recommendationService
             .getRecommendationHistory(recommendationId)
             .map { recommendationQuote -> quoteService.findQuoteById(recommendationQuote.quoteId) }
-            .map(::toQuoteResponse)
+            .map(recommendationResponseMapper::toQuoteResponse)
     }
 
     @Transactional(readOnly = true)
@@ -204,7 +178,7 @@ class RecommendationUseCase(
 
                     RecommendationPeriodItemResponse.from(
                         recommendation = recommendation,
-                        quote = toQuoteResponse(quote),
+                        quote = recommendationResponseMapper.toQuoteResponse(quote),
                     )
                 }
 
@@ -214,68 +188,6 @@ class RecommendationUseCase(
             recommendations = recommendationItems,
         )
     }
-
-    private fun firstRecommendedQuoteId(recommendationId: Long): Long =
-        recommendationService
-            .getRecommendationHistory(recommendationId)
-            .firstOrNull()
-            ?.quoteId
-            ?: throw CustomException(ErrorCode.INVALID_RECOMMENDATION_QUOTE)
-
-    private fun toRecommendationResponse(recommendation: Recommendation): RecommendationResponse {
-        val quote = quoteService.findQuoteById(firstRecommendedQuoteId(recommendation.id))
-
-        return toRecommendationResponse(
-            recommendationId = recommendation.id,
-            quote = quote,
-        )
-    }
-
-    private fun toRecommendationResponse(
-        recommendationId: Long,
-        quote: Quote,
-    ): RecommendationResponse =
-        RecommendationResponse(
-            recommendationId = recommendationId,
-            quote = toQuoteResponse(quote),
-        )
-
-    private fun toRecommendationDetailResponse(recommendation: Recommendation): RecommendationDetailResponse {
-        recommendationService.validateRecommendationCompleted(recommendation)
-
-        val quoteId = recommendation.quoteId ?: throw CustomException(ErrorCode.RECOMMENDATION_NOT_COMPLETED)
-        val quote = quoteService.findQuoteById(quoteId)
-        val (emotionTags, needTag) = toTagDtos(recommendationService.getRecommendationTags(recommendation.id))
-
-        return RecommendationDetailResponse(
-            recommendationId = recommendation.id,
-            quote = toQuoteResponse(quote),
-            emotionRangeId = recommendation.emotionRangeId,
-            emotionTags = emotionTags,
-            needTag = needTag,
-            feelingText = recommendation.feelingText,
-            diaryText = recommendation.diaryText,
-            recommendationDate = recommendation.recommendationDate,
-        )
-    }
-
-    private fun toQuoteResponse(quote: Quote): QuoteResponse {
-        val book = bookService.findBookById(quote.bookId)
-
-        return QuoteResponse.from(
-            quote = quote,
-            book = book,
-        )
-    }
-
-    private fun toTagDtos(recommendationTags: List<RecommendationTag>): Pair<List<TagDto>, TagDto?> {
-        val tagIds = recommendationTags.map { recommendationTag -> recommendationTag.tagId }
-        val (emotionTags, needTag) = emotionService.getEmotionTagsAndNeedTagByIds(tagIds)
-
-        return emotionTags.map(TagDto::from) to needTag?.let(TagDto::from)
-    }
 }
-
-private fun String?.hasText(): Boolean = !isNullOrBlank()
 
 private fun String?.normalizedText(): String? = this?.trim()?.takeIf { text -> text.isNotEmpty() }
